refactor consumers in consumer-group-offset-operation
