+++ internal/consumergroupoffsets/consumer-group-offset-operation.go
@@ -1,8 +1,6 @@
 package consumergroupoffsets
 
 import (
-	"strconv"
-
 	"github.com/deviceinsight/kafkactl/internal/helpers"
 	"golang.org/x/sync/errgroup"
 
@@ -23,14 +21,6 @@ type ResetConsumerGroupOffsetFlags struct {
 	Execute           bool
 	OutputFormat      string
 	allowedGroupState string
-}
-
-type partitionOffsets struct {
-	Partition     int32
-	OldestOffset  int64 `json:"oldestOffset" yaml:"oldestOffset"`
-	NewestOffset  int64 `json:"newestOffset" yaml:"newestOffset"`
-	CurrentOffset int64 `json:"currentOffset" yaml:"currentOffset"`
-	TargetOffset  int64 `json:"targetOffset" yaml:"targetOffset"`
 }
 
 type ConsumerGroupOffsetOperation struct {
@@ -132,7 +122,7 @@ func (operation *ConsumerGroupOffsetOperation) ResetConsumerGroupOffset(flags Re
 	for _, topic := range topics {
 		topicName := topic
 		consumeErrorGroup.Go(func() error {
-			consumer := Consumer{
+			consumer := OffsetResettingConsumer{
 				client:    client,
 				groupName: groupName,
 				topicName: topicName,
@@ -179,32 +169,6 @@ func (operation *ConsumerGroupOffsetOperation) CreateConsumerGroup(flags ResetCo
 type PartitionOffset struct {
 	Offset   int64
 	Metadata string
-}
-
-type SettingOffsetConsumer struct {
-	Topic            string
-	PartitionOffsets map[int32]PartitionOffset
-
-	ready chan struct{}
-}
-
-func (s *SettingOffsetConsumer) Setup(session sarama.ConsumerGroupSession) error {
-	s.ready = make(chan struct{})
-
-	for partition, offset := range s.PartitionOffsets {
-		session.MarkOffset(s.Topic, partition, offset.Offset, offset.Metadata)
-	}
-
-	return nil
-}
-
-func (s *SettingOffsetConsumer) Cleanup(sarama.ConsumerGroupSession) error {
-	close(s.ready)
-	return nil
-}
-
-func (s *SettingOffsetConsumer) ConsumeClaim(sarama.ConsumerGroupSession, sarama.ConsumerGroupClaim) error {
-	return nil
 }
 
 func (operation *ConsumerGroupOffsetOperation) CloneConsumerGroup(srcGroup, targetGroup string) error {
@@ -272,7 +236,7 @@ func (operation *ConsumerGroupOffsetOperation) CloneConsumerGroup(srcGroup, targ
 	for topic, partitionOffsets := range topicPartitionOffsets {
 		topicName, offsets := topic, partitionOffsets
 		consumeErrorGroup.Go(func() error {
-			consumer := SettingOffsetConsumer{
+			consumer := OffsetSettingConsumer{
 				Topic:            topicName,
 				PartitionOffsets: offsets,
 			}
@@ -299,153 +263,6 @@ func (operation *ConsumerGroupOffsetOperation) CloneConsumerGroup(srcGroup, targ
 	output.Infof("consumer-group %s cloned to %s", srcGroup, targetGroup)
 
 	return nil
-}
-
-type Consumer struct {
-	ready     chan bool
-	client    sarama.Client
-	groupName string
-	topicName string
-	flags     ResetConsumerGroupOffsetFlags
-}
-
-func (consumer *Consumer) Setup(session sarama.ConsumerGroupSession) error {
-
-	flags := consumer.flags
-
-	// admin.ListConsumerGroupOffsets(group, nil) can be used to fetch the offsets when
-	// https://github.com/Shopify/sarama/pull/1374 is merged
-	coordinator, err := consumer.client.Coordinator(consumer.groupName)
-	if err != nil {
-		return errors.Wrap(err, "failed to get coordinator")
-	}
-
-	request := &sarama.OffsetFetchRequest{
-		// this will only work starting from version 0.10.2.0
-		Version:       2,
-		ConsumerGroup: consumer.groupName,
-	}
-
-	groupOffsets, err := coordinator.FetchOffset(request)
-	if err != nil {
-		return errors.Wrap(err, "failed to get fetch group offsets")
-	}
-
-	offsets := make([]partitionOffsets, 0)
-
-	if flags.Partition > -1 {
-		offset, err := resetOffset(consumer.client, consumer.topicName, flags.Partition, flags, groupOffsets, session)
-		if err != nil {
-			return err
-		}
-		offsets = append(offsets, offset)
-	} else {
-
-		partitions, err := consumer.client.Partitions(consumer.topicName)
-		if err != nil {
-			return errors.Wrap(err, "failed to list partitions")
-		}
-
-		for _, partition := range partitions {
-			offset, err := resetOffset(consumer.client, consumer.topicName, partition, flags, groupOffsets, session)
-			if err != nil {
-				return err
-			}
-			offsets = append(offsets, offset)
-		}
-	}
-
-	if flags.OutputFormat != "" {
-		if err := output.PrintObject(offsets, flags.OutputFormat); err != nil {
-			return err
-		}
-	} else {
-		tableWriter := output.CreateTableWriter()
-		if err := tableWriter.WriteHeader("PARTITION", "OLDEST_OFFSET", "NEWEST_OFFSET", "CURRENT_OFFSET", "TARGET_OFFSET"); err != nil {
-			return err
-		}
-		for _, o := range offsets {
-			if err := tableWriter.Write(strconv.FormatInt(int64(o.Partition), 10),
-				strconv.FormatInt(o.OldestOffset, 10), strconv.FormatInt(o.NewestOffset, 10),
-				strconv.FormatInt(o.CurrentOffset, 10), strconv.FormatInt(o.TargetOffset, 10)); err != nil {
-				return err
-			}
-		}
-		if err := tableWriter.Flush(); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (consumer *Consumer) Cleanup(sarama.ConsumerGroupSession) error {
-	close(consumer.ready)
-	return nil
-}
-
-func (consumer *Consumer) ConsumeClaim(sarama.ConsumerGroupSession, sarama.ConsumerGroupClaim) error {
-	return nil
-}
-
-func getPartitionOffsets(client sarama.Client, topic string, partition int32, flags ResetConsumerGroupOffsetFlags) (partitionOffsets, error) {
-
-	var err error
-	offsets := partitionOffsets{Partition: partition}
-
-	if offsets.OldestOffset, err = client.GetOffset(topic, partition, sarama.OffsetOldest); err != nil {
-		return offsets, errors.Errorf("failed to get offset for topic %s Partition %d: %v", topic, partition, err)
-	}
-
-	if offsets.NewestOffset, err = client.GetOffset(topic, partition, sarama.OffsetNewest); err != nil {
-		return offsets, errors.Errorf("failed to get offset for topic %s Partition %d: %v", topic, partition, err)
-	}
-
-	if flags.Offset > -1 {
-		if flags.Offset < offsets.OldestOffset {
-			return offsets, errors.Errorf("cannot set offset for Partition %d: offset (%d) < oldest offset (%d)", partition, flags.Offset, offsets.OldestOffset)
-		} else if flags.Offset > offsets.NewestOffset {
-			return offsets, errors.Errorf("cannot set offset for Partition %d: offset (%d) > newest offset (%d)", partition, flags.Offset, offsets.NewestOffset)
-		} else {
-			offsets.TargetOffset = flags.Offset
-		}
-	} else {
-		if flags.OldestOffset {
-			offsets.TargetOffset = offsets.OldestOffset
-		} else if flags.NewestOffset {
-			offsets.TargetOffset = offsets.NewestOffset
-		} else {
-			return offsets, errors.New("either offset,oldest,newest parameter needs to be specified")
-		}
-	}
-
-	return offsets, nil
-}
-
-func resetOffset(client sarama.Client, topic string, partition int32, flags ResetConsumerGroupOffsetFlags, groupOffsets *sarama.OffsetFetchResponse, session sarama.ConsumerGroupSession) (partitionOffsets, error) {
-	offset, err := getPartitionOffsets(client, topic, partition, flags)
-	if err != nil {
-		return offset, err
-	}
-
-	offset.CurrentOffset = getGroupOffset(groupOffsets, topic, partition)
-
-	if flags.Execute {
-		if offset.TargetOffset > offset.CurrentOffset {
-			session.MarkOffset(topic, partition, offset.TargetOffset, "")
-		} else if offset.TargetOffset < offset.CurrentOffset {
-			session.ResetOffset(topic, partition, offset.TargetOffset, "")
-		}
-	}
-
-	return offset, nil
-}
-
-func getGroupOffset(offsetFetchResponse *sarama.OffsetFetchResponse, topic string, partition int32) int64 {
-	block := offsetFetchResponse.Blocks[topic][partition]
-	if block != nil {
-		return block.Offset
-	}
-	return -1
 }
 
 type DeleteConsumerGroupOffsetFlags struct {
